+++ src/main/java/uni/quindio/eduquizsolutions/repositories/ContenidosRepo.java
@@ -6,6 +6,6 @@ import org.springframework.stereotype.Repository;
 import uni.quindio.eduquizsolutions.entities.Contenidos;
 
 @Repository
-public interface ContenidosRepo extends JpaRepository<Contenidos, Integer> {
+public interface ContenidosRepo extends JpaRepository<Contenidos, Long> {
 
 }
