+++ src/main/java/uni/quindio/eduquizsolutions/repositories/DiaRepo.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import uni.quindio.eduquizsolutions.entities.Dias;
 import org.springframework.stereotype.Repository;
 @Repository
-public interface DiaRepo extends JpaRepository<Dias, Integer> {
+public interface DiaRepo extends JpaRepository<Dias, Long> {
 
 }
