+++ src/main/java/uni/quindio/eduquizsolutions/repositories/HorariosRepo.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import uni.quindio.eduquizsolutions.entities.Horarios;
 import org.springframework.stereotype.Repository;
 @Repository
-public interface HorariosRepo extends JpaRepository<Horarios, Integer> {
+public interface HorariosRepo extends JpaRepository<Horarios, Long> {
 
 }
