+++ src/main/java/uni/quindio/eduquizsolutions/repositories/InstitucionesRepo.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import uni.quindio.eduquizsolutions.entities.Instituciones;
 import org.springframework.stereotype.Repository;
 @Repository
-public interface InstitucionesRepo extends JpaRepository<Instituciones, Integer> {
+public interface InstitucionesRepo extends JpaRepository<Instituciones, Long> {
 
 }
