+++ src/main/java/uni/quindio/eduquizsolutions/repositories/PreguntasRepo.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import uni.quindio.eduquizsolutions.entities.Preguntas;
 import org.springframework.stereotype.Repository;
 @Repository
-public interface PreguntasRepo extends JpaRepository<Preguntas, Integer> {
+public interface PreguntasRepo extends JpaRepository<Preguntas, Long> {
 
 }
