+++ src/main/java/uni/quindio/eduquizsolutions/repositories/RespuestasEstudianteRepo.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import uni.quindio.eduquizsolutions.entities.RespuestasEstudiantes;
 import org.springframework.stereotype.Repository;
 @Repository
-public interface RespuestasEstudianteRepo extends JpaRepository<RespuestasEstudiantes, Integer> {
+public interface RespuestasEstudianteRepo extends JpaRepository<RespuestasEstudiantes, Long> {
 
 }
