+++ src/main/java/uni/quindio/eduquizsolutions/repositories/RespuestasRepo.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import uni.quindio.eduquizsolutions.entities.Respuestas;
 import org.springframework.stereotype.Repository;
 @Repository
-public interface RespuestasRepo extends JpaRepository<Respuestas, Integer> {
+public interface RespuestasRepo extends JpaRepository<Respuestas, Long> {
 
 }
