+++ src/main/java/uni/quindio/eduquizsolutions/repositories/TemasRepo.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import uni.quindio.eduquizsolutions.entities.Temas;
 import org.springframework.stereotype.Repository;
 @Repository
-public interface TemasRepo extends JpaRepository<Temas, Integer> {
+public interface TemasRepo extends JpaRepository<Temas, Long> {
 
 }
