+++ src/main/java/uni/quindio/eduquizsolutions/repositories/TiposPreguntasRepo.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import uni.quindio.eduquizsolutions.entities.TiposPreguntas;
 import org.springframework.stereotype.Repository;
 @Repository
-public interface TiposPreguntasRepo extends JpaRepository<TiposPreguntas, Integer> {
+public interface TiposPreguntasRepo extends JpaRepository<TiposPreguntas, Long> {
 
 }
